requirements.txt
# Latest version of plotly can be found at https://github.com/plotly/plotly.py/releases .
# Version is hardcoded to ensure uniformity no matter the platform.
plotly==5.24.1
polars==1.30.0
dash~=2.18
pytest==8.4.0
pylint==3.3.7
black==25.1.0
black[jupyter]
pyodbc
numpy==2.0.2
gov-uk-dashboards~=24.1.0
uk-gov-dash-components~=1.29.4
dash[testing]
chromedriver-autoinstaller
playwright
pytest-playwright
pixelmatch
pillow~=11.2
snakeviz
werkzeug
python-dateutil>=2.8.0
pydantic
sqlalchemy==2.0.41
azure-storage-blob
sentry-sdk[flask]
python-dotenv
bleach